Add Initial Spec for User Model
This commit adds test to ensure that the user model is responding
to its attributes. It creates a user factory in FactoryGirl and
uses an explicit subject at the top of the test's describe block and
the `is_expected.to` method in RSpec, which minimizes some of the typing
around tests for simple conditions.